Move inside AutoForeground for better correlation
