+++ libs/utils/WordPressUtils/src/main/java/org/wordpress/android/util/AutoForeground.java
@@ -2,7 +2,10 @@ package org.wordpress.android.util;
 
 import android.app.Notification;
 import android.app.Service;
+import android.content.ComponentName;
+import android.content.Context;
 import android.content.Intent;
+import android.content.ServiceConnection;
 import android.os.Binder;
 import android.os.IBinder;
 import android.support.annotation.CallSuper;
@@ -16,6 +19,33 @@ public abstract class AutoForeground<EventClass> extends Service {
     public static final int NOTIFICATION_ID_PROGRESS = 1;
     public static final int NOTIFICATION_ID_SUCCESS = 2;
     public static final int NOTIFICATION_ID_FAILURE = 3;
+
+    public static class ServiceEventConnection {
+        private final ServiceConnection mServiceConnection;
+
+        public ServiceEventConnection(Context context, Class<? extends AutoForeground> clazz, Object client) {
+            EventBus.getDefault().register(client);
+
+            mServiceConnection = new ServiceConnection() {
+                @Override
+                public void onServiceConnected(ComponentName componentName, IBinder iBinder) {
+                    // nothing here
+                }
+
+                @Override
+                public void onServiceDisconnected(ComponentName componentName) {
+                    // nothing here
+                }
+            };
+
+            context.bindService(new Intent(context, clazz), mServiceConnection, Context.BIND_AUTO_CREATE);
+        }
+
+        public void disconnect(Context context, Object client) {
+            context.unbindService(mServiceConnection);
+            EventBus.getDefault().unregister(client);
+        }
+    }
 
     private class LocalBinder extends Binder {}
